requirements.txt
google-cloud-bigquery
google-auth
pandas
python-dotenv
jupyter
pyyaml
